We now do a setup for the user, not for the subproject. User is defined by username.
We also default to taking the current directory as the base ChildHH
directory, meaning that we expect you do cd to your copy of the base
project code.  You can specify that we should use the Box copy of
the code, but that's not defined right now.

We no longer look for a default_project.txt file.