partner_change variables
pchangehc was edited so that a transition from cohabitation to marriage is not counted as a partner change.

idpartner was using other variables to identify partner change. These other variables would have still counted a move from cohabitation to marriage as a change. Edited idpartner.do so that it uses the variables created by pchangehc.do instead.