Don’t carry relationship source into next transitive relationships
Get rid of old relationship sources when generating a new round of
relationships.  But it may be interesting to know where the
relationships came from when generating new ones, so create a
relsource1 and relsource2 to go with relationship1 and relationship2.
Right now I’m not doing anything with them.