Fix an "off by one" error where we should only copy the next wave's idea of who's present in the gap if this is not the last wave of the gap.  In other words, for the last wave of the gap, don't copy the previous wave's record of who's present because we don't care!  We're starting the computation of a new gap.
Also, make sure we don't do anything for non-existent members of the
list of those in the previous HH.  We have to run through the loop past
the end of most HH lists because we have to loop enough to cover the longest
list.  Once we run out of members in an HH, don't do anything to modify
the state.