Added insample a more explicit definition of sample
previous code goes to some effort to code hh_change==1 even if missing in this wave if someone appears in next wave in a household with people who appear in this wave. We want to be sure to keep these observations even. These cases are missing on ERRP in the wave without interview.

Also revised results file to incorporate changes to code over the past couple of months. The results are still similar to what we presented in the previous version of the paper.